Add tool to re-encode blobs to newest protocol
This tool uses the current version of Sereal with the currently best
compression settings to re-encode the blobs in the per-version module
files.

This commit also updates the settings in the regular generation logic
such that it generates with the same updated settings.

This drops the "canonical" flag from the generation since it doesn't
really add value here.